refactor: Move field centric drive functionality to "Drivetrain" class
This commit should move the field centric drive functionality to the drivetrain class. This should improve code readability and make it easier to work with.

Additionally,  the Drivetrain's starting position is no longer a parameter for its "init()" method, and is now contained within it's constructor.

The RobotHardware class has had a few alteration that are aimed at improving performance.