feat!: enhance command system functionality
This commit adds several features that enhance the functionality of the command system. Firstly, when scheduled, commands now have the ability to stop commands that use the same subsystem as the newly running command based on which command has a higher priority. This prevents unexpected behavior from having two commands use the same subsystem at once. Secondly, event commands have been added! These are scheduled with the EventManager and take in an event listener to determine when each command should be run.

BREAKING CHANGES:
- Sequential command groups are, as of present, deprecated. This will be fixed at a later date.